Clean shutdown implementation Storage queue fix

+++ eu.dariolucia.reatmetric.ui/src/main/java/eu/dariolucia/reatmetric/ui/controller/DateTimePickerWidgetController.java
@@ -17,14 +17,6 @@
 
 package eu.dariolucia.reatmetric.ui.controller;
 
-import java.net.URL;
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.time.ZoneOffset;
-import java.util.ResourceBundle;
-
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
@@ -32,6 +24,10 @@ import javafx.scene.control.Button;
 import javafx.scene.control.DatePicker;
 import javafx.scene.control.Spinner;
 import javafx.scene.control.SpinnerValueFactory;
+
+import java.net.URL;
+import java.time.*;
+import java.util.ResourceBundle;
 
 /**
  * FXML Controller class
@@ -50,11 +46,12 @@ public class DateTimePickerWidgetController implements Initializable {
     private DatePicker datePicker;
     @FXML
     private Button selectBtn;
-    
+
     private Runnable actionAfterSelection;
-    
+
     // The result of the selection
     private Instant selectedTime = null;
+
     /**
      * Initializes the controller class.
      */
@@ -64,20 +61,20 @@ public class DateTimePickerWidgetController implements Initializable {
         this.minuteSpn.setValueFactory(new SpinnerValueFactory.IntegerSpinnerValueFactory(0, 59, 0));
         this.secondSpn.setValueFactory(new SpinnerValueFactory.IntegerSpinnerValueFactory(0, 59, 0));
         this.datePicker.setValue(LocalDate.now());
-    }  
-    
+    }
+
     @FXML
     private void selectButtonPressed(ActionEvent e) {
         this.selectedTime = deriveFromWidgets();
-        if(this.actionAfterSelection != null) {
+        if (this.actionAfterSelection != null) {
             this.actionAfterSelection.run();
         }
     }
-  
+
     public void setActionAfterSelection(Runnable r) {
         this.actionAfterSelection = r;
     }
-    
+
     public void setSelectedTime(Instant t) {
         this.selectedTime = t;
         updateWidgets();
@@ -86,7 +83,7 @@ public class DateTimePickerWidgetController implements Initializable {
     public Instant getSelectedTime() {
         return this.selectedTime;
     }
-    
+
     private void updateWidgets() {
         LocalDateTime ldt = LocalDateTime.ofInstant(this.selectedTime, ZoneOffset.UTC);
         this.hourSpn.getValueFactory().setValue(ldt.getHour());
@@ -96,7 +93,7 @@ public class DateTimePickerWidgetController implements Initializable {
     }
 
     private Instant deriveFromWidgets() {
-        LocalDateTime ldt = LocalDateTime.of(this.datePicker.getValue(), LocalTime.of((int) this.hourSpn.getValue(),(int) this.minuteSpn.getValue(),(int) this.secondSpn.getValue()));
+        LocalDateTime ldt = LocalDateTime.of(this.datePicker.getValue(), LocalTime.of(this.hourSpn.getValue(), this.minuteSpn.getValue(), this.secondSpn.getValue()));
         return ldt.toInstant(ZoneOffset.UTC);
     }
 }
